Move TimeFormat helper class to support package
