+++ app/src/main/java/com/mobilegenomics/f5n/activity/ConfirmationActivity.java
@@ -19,12 +19,10 @@ import android.widget.TextView;
 import android.widget.Toast;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
-
 import com.mobilegenomics.f5n.GUIConfiguration;
 import com.mobilegenomics.f5n.R;
-import com.mobilegenomics.f5n.core.TimeFormat;
 import com.mobilegenomics.f5n.core.PipelineComponent;
-
+import com.mobilegenomics.f5n.support.TimeFormat;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileOutputStream;
